TXNCXX-180 Fit performer query support
Motivation
==========
Now that queries are in the library, lets test them in fit.

Modification
============
The tests pointed out a number of small issues, which are addressed
here.  Mostly in error handling, adding hooks (though before/after
query not implemented yet).

Added a json string test to the transaction_context tests, to verify an
issue was not seen here that was seen elsewhere.

Results
=======
Query tests on fit run, with about 60-ish errors

Change-Id: I094fa420cedf82efe17cfe1cbcae798502756289
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/169346
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>