TXNCXX-193: Send all queries in txn to same node.
Motivation
==========
Once you call begin_work, you have to send all subsequent queries to the
node the begin_work went to, for that attempt.

Modification
============
Anticipating this, we already keep track of the node that responded to
the begin_work call.   So, the wrap_query just pops that in there, as
long as it is not a begin_work call.

Removed the only place we use the error_class inside the
transaction_op_failed exception.   We can remove it safely from the
exception entirely, now.   But that will be in a separate commit, just
wanted this in now, to ensure we don't use it on the fit side of things

Updated the test environment to ping until kv and query are available, which
should prevent intermittent errors caused by racing the availability of
the query service.  Also, made the txn timeout 5 seconds now, to account
for query durations that can be much longer than kv.

Results
=======
Tests pass against cluster with multiple query nodes.

Change-Id: Ia771514d0abdb0ab8e0ed36c551c5a03e45e2789
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/169360
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>