TXNCXX-198: Fit Performer test failures
Motivation
==========
We should get to passing all 3000 or so tests.

Modification
============
Not handling a query_exception in the transaction lambda should result
in a transaction_operation_failed exception.   In some cases, we can
take the specific exception type, and map it to an external_exception as
the cause (for FIT tests).

Rollback and Commit should check for is_done before running - handling
cases where finalize is waiting, and an explicit commit or rollback is
called.  We will eliminate this from the attempt_context interface at
some point, but the wrappers may still want to invoke it directly, or
via the transaction_context.

Added a number of new external_exception members, mapping the
PARSING_FAILURE accordingly.  More to do here with other causes, but
these were the most obvious ones preventing a number of query tests from
passing.

The get_optional, when in query mode, needs to not return an exception
when the document isn't found.  Sprinkled in a bit of code to do that.

Added a few unit tests for all this as well

Results
=======
Passing unit tests, less fit failures.  No complaints during unit tests
from tsan or asan.  More to come.

Change-Id: I5adfbc8118ab84af656cd48efbbd05196dd62518
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/169682
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>