TXNCXX-196: Add support for BF-CBD-3838 TXNCXX-197: Setting ATR to Committed is not inserting "p" field
BF-CBD-3838 corrects the handling of FAIL_AMBIGUOUS
during setting the ATR to COMMITTED.  It involves
retrying the write, and rereading the ATR to see what
state is is now.

In passing, also correcting a bug where the "p" field
is not being inserted at COMMITTED point.  This is
important as it prevents collision with another actor
trying to modify this entry - such as cleanup, or in the
future rage.

Change-Id: I59a48d58b654cc08ea0ca78dd0aae7de6ddaa05b
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/169729
Tested-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>