TXNCXX-170  Serializing staged documents
Motivation
==========
Fit performer failures with current fix for this bug.

Modification
=============
Seems I forgot there are 2 ways to get the transaction_get_result, and
only fixed one of them, in terms of getting the staged atr scope, etc...

This properly addresses that - it leads to crashes in cleanup (not
sufficiently tested in unit tests currently).

Results
=======
The fit test failure (crash) is fixed.

Change-Id: I69e79d077467abc24ca88b179fa8ff754a898a6f
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/169772
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>