TXNCXX-211 and TXNCXX-212 - Scan Consistency, metrics, etc...
Motivation
===========
Expose metrics, to ease integration issues.   While at it, expose the
max_parallelism (which was forgotten), and insure the configured
scan_consistency is used.

Modification
============
The scan_consistency should not have had a default - since there is one
in the config.  That has been removed.  Added metrics, and
max_parallelism as well.

Results
=======
passing tests.

Change-Id: Ia8c80d3cfb503cdda74341bb886c1d6652918fe0
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/169950
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>