TXNCXX-180 Fit performer query support
Motivation
==========
Original fix had forgotten to add the "QU" to the supported extensions.
Also, there seems to have been a regression - the cause for
query_exception was not being set in the constructor any more.

Modification
============
Added "QU", updated query_exception constructor to initialize the cause.

Result
======
Fit, which now uses the supported extensions in the forward_compat to
signal what the library supports, knows to run the query tests.  Also,
it now properly sees the cause for query_exceptions, which are checked
in a number of query tests.

Change-Id: Id485dd7d50347cf3a9351d01b1656af988796fe3
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/169958
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>