TXNCXX-225 ExtCustomMetadataCollection
Motivation
==========
Support for all the 2.1 protocol extensions.

Modification
============
Added a metadata_collection member to the transaction_config.  Sprinkled
the use of it when getting atrs around as needed.

While there, noticed an issue with the atr_commit retrying recursively,
which can blow up the stack, so used retry_op<void> instead, etc...

Results
=======
Fit tests are happy with this, locally.

Change-Id: I0d2b788ad138e1d32138ec4255ca3ff2dfabf365
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/170115
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>