Revert "Removing the delay parameter from creating a staged insert"
This reverts commit 938ae7911bf01f32305df691332cd5937249adfb.

Reason for revert: reverting as this creates stack overflow when (in fit) having the before_staged_insert raise a FAIL_TRANSIENT until timeout.  This was why it was there - a way to backoff to keep this from having too deep a stack.   Ideally we don't call recursively in a tight loop -- and in practice this wouldn't happen outside fit in such a tight loop to create a super deep stack.   Still worth eliminating the recursion, eventually.  An exponential backoff is fine for now.

Change-Id: I81ab16fde73dbbd3530d50cc4910ca76e827b5dd
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/169550
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>