TXNCXX-234: Lost cleanup is not inserting "p" field TXNCXX-233: Lost cleanup deletes all ATR entries if it deletes one TXNCXX-232: Cleanup takes longer than the specified window
Fixing three lost cleanup bugs.

TXNCXX-233 means that if one ATR entry is cleaned up, then all
entries in that ATR were removed.

TXNCXX-232 means that the cleanup window was not being followed very
closely.  This adds an algo that dynamically adjusts the 'time budget' for
reading each ATR, based on how long is left in the cleanup
window and how many ATRs remain.  As long as the total time
spent reading the ATRs at the KV level, and subsequently cleaning
them up, does not exceed the cleanup window, that window should
now be followed closely.

TXNCXX-234 means that the "p" field was not being inserted while
removing a PENDING entry.  This could lead to it conflicting
with the main algo in the unlikely event that is still running
even though the transaction has expired.

Also re-adding "QU" that was accidentally removed from extensions.

Change-Id: Ib5af7c03e0e1fa42aba6b0ee4e15f6b213c99be3
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/170387
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Tested-by: David Kelly <davidmichaelkelly@gmail.com>