TXNCXX-237: Segfault with unexpected error in query
Motivation
==========
Segfaults are bad, mkay.

Modification
============
Check for no errors in meta, and map to a transaction_operation_failed
exception.  Though you can trigger this with CXXCBC-131, there are other
errors which may have lead to this as well.

Results
=======
Using "localhost" for the connection string in the tests, you can see no
segfault now.   But the underlying issue in the client is still there,
so queries don't work.  When CXXCBC-131 is fixed, that will no longer be
an issue.

Change-Id: Ie9a622a1321d853398a6484ad398df194fba41f5
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/170586
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Build Bot <build@couchbase.com>