TXNCXX-183 Crash on M1
Motivation
==========
Experiment in adjusting how we open buckets at startup, in hopes of
seeng a change in these crashes.

Modification
============
It seems we could be interrupted, notice that the count = 0 and return
from get_and_open_buckets, after the last open_bucket handler has
decremented the count to 0 but before it calls notify_all on the
condition variable. Rather than extent the lock needlessly, lets just
make that a shared_ptr.   Added tests for this while at it.

Also noticed we created the couchbase console logger 2 times.   That's
why we were not seeing the couchbase-cxx-client logging in tests.
Simple fix.

Results
=======
Works fine locally, but it did before too.   Lets see how jenkins feels
about it on m1.

Change-Id: If9cbb65940cebb0a014b010961461153f6224de3
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/170910
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>