TXNCXX-244 Very intermittent crash opening buckets
Motivation
==========
Crashing is bad.

Modification
============
When opening buckets, the last one calls notify() on a
condition_variable which is being waited upon.  If that waiting function
then completes before the line after the notify, the mutex (passed by
reference into the lambda) is destroyed and the subsequent unlock will
segfault.   Switched to a shared_ptr.

Result
======
Cannot reproduce it in the python tests (where I did see it often).

Change-Id: I017d6b96f94915e54a59c0e7ed99302c93676474
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/173253
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>