TXNCXX-252: Support custom logger sink
Motivation
==========
For wrappers (python, node, etc...) to be able to use their particular
logging ecosystem, we need to allow them to create a custom spdlog
sink that does that handoff.  Additionally, any c++ users should be able
to do the same - log to a file, syslog, etc...

Modification
===========
Added new function create_loggers which takes a logger level and a
shared_ptr to the desired sink.  This replaces the existing sink(s) for
all the loggers with the one passed in, if any, and ajusts the level
accordinly.

Also added some tests of both the default logging and the use of a
custom sink.

Results
=======
The tests all pass.

Change-Id: I65dc8d490700c61bb43ee02b8ce4cb646fac4d82
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/176257
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>