TXNCXX-250 - Open custom metadata bucket if one specified
Motivation
==========
There will be a race with the cleanup thread, which opens all the
buckets when starting up.  If that thread hasn't opened the custom
metadata bucket before a transaction is run, then the transaction will
fail.

Modification
============
We do a blocking open on the bucket when transactions object is created.
The client caches the buckets, so if already open, this does no i/o.
Now, if the collection or scope are not there, the transaction fails
after retrying until timeout.  We could check for their presence, but
this still races if not all nodes have updated to have it yet, etc...
This seems like a reasonable balance between efficiency and convenience.

Added a test that uses custom metadata collections, along with 3 that
test the failure modes.

Results
=======
Tests pass, however the scripts on Jenkins need to be updated to create
the second bucket.

Change-Id: I1cf75e303007a4e1b8c6d6c91cd4c4e59f6dca30
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/176393
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>