TXNCXX-261 Update client after initial api split
Motivation
==========
The cxx client is growing its public interface.  We actually use the
internal api, and now that we are growing a public api, the namespacing
and corresponding includes will change.   Lets keep everything in sync
as this proceeds.

Modification
============
Updated client, then corresponding namespaces and includes for those
things that changed.  Mostly the operations, cluster (use the core
and not the couchbase one), and the couchbase::errc namespace with
corresponding changes to names of the error_codes.

Results
======
Build and tests pass.

Change-Id: I45f73d61c57b5efef732298c038e0ae199e520b3
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/178004
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>