TXNCXX-248 Correct timeout for query txn operations.
Motivation
==========
The durable kv timeout was not exposed at some point, and I defaulted to
the kv timeout if not set in the config(s).   Now it is exposed, we
should use it.

Modification
============
In begin_work, we need to set it correctly.   Simple change.

While at it, made sure we didn't set the timeout incorrectly elsewhere.
Seems we only have to set it in begin work, so that was the only place I
needed the correct default.

Results
=======
All tests pass locally.   Didn't add a test for this, as it isn't easy
to test reliably.   Ideally, we'd have a mock server that we could ask
to delay before responding, so we could test the client's behavior in a
repeatable fashion.  But AFAIK, we can't so far.

Change-Id: Icf10e709de3d6ba5d85e18c6a4a888600734e5a0
Reviewed-on: https://review.couchbase.org/c/couchbase-transactions-cxx/+/179683
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>